requirements.txt
altair==4.1.0
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
matplotlib==3.5.0
pandas==1.3.4
pytz==2021.3